% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ucl_scales.R
\name{scale_colour_ucl}
\alias{scale_colour_ucl}
\alias{scale_color_ucl}
\alias{scale_fill_ucl}
\title{UCL colour scales for colour and fill aesthetics}
\usage{
scale_colour_ucl(palette = "blues", discrete = TRUE, reverse = FALSE, ...)

scale_color_ucl(...)

scale_fill_ucl(palette = "blues", discrete = TRUE, reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of one of the palettes}

\item{discrete}{Should the palette represent discrete values or continuous
values?}

\item{reverse}{Should the order of values in the palette be reversed?}

\item{...}{Further arguments passed to \code{scale_colour_manual} for
discrete palettes and \code{scale_colour_gradientn} for continuous
palettes.}
}
\description{
UCL colour scales for colour and fill aesthetics
}
\section{Functions}{
\itemize{
\item \code{scale_color_ucl()}: UCL colour scales for colour and fill aesthetics

\item \code{scale_fill_ucl()}: UCL colour scales for colour and fill aesthetics

}}
