% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ucl_colours.R
\name{ucl_colours}
\alias{ucl_colours}
\alias{ucl_palette}
\title{Return colour values from the UCL colour scheme}
\usage{
ucl_colours(...)

ucl_palette(palette = NULL)
}
\arguments{
\item{...}{Character values giving the names of one or more colours from the
UCL colour scheme. If no values are supplied, all possible colours are
returned.}

\item{palette}{Character name of a palette.}
}
\value{
A named vector of hex colour values.

Calling \code{ucl_palette} with no arguments will return the names of all the
available palettes. More details are available on the \href{https://www.ucl.ac.uk/cam/brand/guidelines/colour}{UCL style guide page on colour palettes}.
}
\description{
Return colour values from the UCL colour scheme
}
\section{Functions}{
\itemize{
\item \code{ucl_palette()}: Return colour values from the UCL colour scheme

}}
